I had understood the measure is empty test wrong. In fact a measure is not empty when there are only rests. It is silent. Silent measures are valid in PS and should not be skipped.
An empty measure is a measure with no notes at all not even rest(s).
This is invalid XML I do not think Musescore while exporting will ever
generate this situation but it is good as an additional safety net.

- Renamed the current measureIsEmpty() function to measureIsSilent()
this function is not used anywhere but kept for future usage.

- Fixed the measureIsEmpty() function to reflect this state and skip
that measure.